Start
This is the start of the app.
I'm on a bad network, so it may take a while for something to show up.
But I'm trying to make this app so it will tell you when you get into a
3G+ zone. (I ran into a problem when I was riding for 2 hours and wanted
to watch Netflix. This could've made my trip a lot better)